pyproject.toml
[tool.poetry]
name = "easy-wrk"
version = "0.1.0"
description = ""
authors = ["riag <riag@163.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.25.1"
attrs = "^20.3.0"
cattrs = "^1.1.2"
Jinja2 = "^2.11.2"
tomlkit = "^0.7.0"
tabulate = "^0.8.7"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
